[CODE]: optim for factories
