+++ arch/core/common/PipelineStage.scala
