--- arch/core/csr/abstract.scala
@@ -1,6 +1,0 @@
-package arch.core.csr
-
-trait CsrUtilities {
-  def cmdWidth: Int
-  def addrWidth: Int
-}
