+++ arch/core/decoder/Components.scala
@@ -1,12 +1,12 @@
 package arch.core.decoder
 
-import arch.configs._
 import arch.core.bru._
 import arch.core.imm._
 import arch.core.alu._
 import arch.core.lsu._
 import arch.core.regfile._
 import arch.core.csr._
+import arch.configs._
 import chisel3._
 import chisel3.util.BitPat
 
@@ -46,9 +46,15 @@ class DecodedOutput(implicit p: Parameters) extends Bundle {
   val csr_cmd = UInt(csr_utils.cmdWidth.W)
 }
 
-trait DecoderUtilities {
+trait DecoderUtilities extends Utilities {
   def default: List[BitPat]
   def bubble: BitPat
   def decode(instr: UInt): DecodedOutput
   def table: Array[(BitPat, List[BitPat])]
+}
+
+object DecoderUtilitiesFactory extends UtilitiesFactory[DecoderUtilities]("Decoder")
+
+object DecoderInit {
+  val rv32iUtils = RV32IDecoderUtilities
 }
