--- arch/core/decoder/factory.scala
@@ -1,36 +1,0 @@
-package arch.core.decoder
-
-import scala.collection.mutable
-
-object DecoderUtilitiesFactory {
-  private val registry = mutable.Map[String, DecoderUtilities]()
-
-  def register(name: String, sigs: DecoderUtilities): Unit =
-    registry(name.toLowerCase) = sigs
-
-  def get(name: String): Option[DecoderUtilities] =
-    registry.get(name.toLowerCase)
-
-  def getOrElse(name: String, default: DecoderUtilities): DecoderUtilities =
-    registry.getOrElse(name.toLowerCase, default)
-
-  def getOrThrow(name: String): DecoderUtilities =
-    registry.getOrElse(
-      name.toLowerCase,
-      throw new NoSuchElementException(s"DecoderUtilities for ISA '$name' not found")
-    )
-
-  def listAvailable(): Seq[String] = registry.keys.toSeq.sorted
-
-  def contains(name: String): Boolean = registry.contains(name.toLowerCase)
-}
-
-trait RegisteredDecoderUtilities {
-  def isaName: String
-  def utils: DecoderUtilities
-  DecoderUtilitiesFactory.register(isaName, utils)
-}
-
-object DecoderInit {
-  val rv32iUtils = RV32IDecoderUtilities
-}
