--- arch/core/imm/abstract.scala
@@ -1,9 +1,0 @@
-package arch.core.imm
-
-import chisel3._
-
-trait ImmUtilities {
-  def immTypeWidth: Int
-  def createBundle: UInt
-  def genImm(instr: UInt, immType: UInt): UInt
-}
