--- arch/core/imm/factory.scala
@@ -1,36 +1,0 @@
-package arch.core.imm
-
-import scala.collection.mutable
-
-object ImmUtilitiesFactory {
-  private val registry = mutable.Map[String, ImmUtilities]()
-
-  def register(name: String, sigs: ImmUtilities): Unit =
-    registry(name.toLowerCase) = sigs
-
-  def get(name: String): Option[ImmUtilities] =
-    registry.get(name.toLowerCase)
-
-  def getOrElse(name: String, default: ImmUtilities): ImmUtilities =
-    registry.getOrElse(name.toLowerCase, default)
-
-  def getOrThrow(name: String): ImmUtilities =
-    registry.getOrElse(
-      name.toLowerCase,
-      throw new NoSuchElementException(s"ImmUtilities for ISA '$name' not found")
-    )
-
-  def listAvailable(): Seq[String] = registry.keys.toSeq.sorted
-
-  def contains(name: String): Boolean = registry.contains(name.toLowerCase)
-}
-
-trait RegisteredImmUtilities {
-  def isaName: String
-  def utils: ImmUtilities
-  ImmUtilitiesFactory.register(isaName, utils)
-}
-
-object ImmInit {
-  val rv32iUtils = RV32IImmUtilities
-}
