+++ arch/core/lsu/Components.scala
@@ -1,8 +1,9 @@
 package arch.core.lsu
 
+import arch.configs._
 import chisel3._
 
-trait LsuUtilities {
+trait LsuUtilities extends Utilities {
   def cmdWidth: Int
   def strb(cmd: UInt): UInt
 
@@ -14,4 +15,10 @@ trait LsuUtilities {
   def isWrite(cmd: UInt): Bool
   def isMemRead(is_mem: Bool, cmd: UInt): Bool
   def isMemWrite(is_mem: Bool, cmd: UInt): Bool
+}
+
+object LsuUtilitiesFactory extends UtilitiesFactory[LsuUtilities]("LSU")
+
+object LsuInit {
+  val rv32iUtils = RV32ILsuUtilities
 }
