+++ arch/core/lsu/variants/rv32i.scala
@@ -1,6 +1,7 @@
 package arch.core.lsu
 
 import arch.core.common.Consts
+import arch.configs._
 import chisel3._
 import chisel3.util._
 
@@ -18,28 +19,28 @@ trait RV32ILsuConsts extends Consts {
   def M_LHU = BitPat("b1101")
 }
 
-class RV32ILsuUtilitiesImpl extends LsuUtilities with RV32ILsuConsts {
-  def cmdWidth: Int   = SZ_M
-  def strb(cmd: UInt) =
-    MuxLookup(cmd(1, 0), "b0000".U(4.W))(
-      Seq(
-        "b00".U -> "b0001".U(4.W),
-        "b01".U -> "b0011".U(4.W),
-        "b10".U -> "b1111".U(4.W)
+object RV32ILsuUtilities extends RegisteredUtilities[LsuUtilities] with RV32ILsuConsts {
+  override def utils: LsuUtilities = new LsuUtilities {
+    override def name: String = "rv32i"
+
+    override def cmdWidth: Int                             = SZ_M
+    override def strb(cmd: UInt)                           =
+      MuxLookup(cmd(1, 0), "b0000".U(4.W))(
+        Seq(
+          "b00".U -> "b0001".U(4.W),
+          "b01".U -> "b0011".U(4.W),
+          "b10".U -> "b1111".U(4.W)
+        )
       )
-    )
+    override def isByte(cmd: UInt): Bool                   = cmd(1, 0) === "b00".U
+    override def isHalf(cmd: UInt): Bool                   = cmd(1, 0) === "b01".U
+    override def isWord(cmd: UInt): Bool                   = cmd(1, 0) === "b10".U
+    override def isUnsigned(cmd: UInt): Bool               = cmd(2)
+    override def isRead(cmd: UInt): Bool                   = cmd(3)
+    override def isWrite(cmd: UInt): Bool                  = !cmd(3)
+    override def isMemRead(is_mem: Bool, cmd: UInt): Bool  = is_mem && cmd(3)
+    override def isMemWrite(is_mem: Bool, cmd: UInt): Bool = is_mem && !cmd(3)
+  }
 
-  def isByte(cmd: UInt): Bool                   = cmd(1, 0) === "b00".U
-  def isHalf(cmd: UInt): Bool                   = cmd(1, 0) === "b01".U
-  def isWord(cmd: UInt): Bool                   = cmd(1, 0) === "b10".U
-  def isUnsigned(cmd: UInt): Bool               = cmd(2)
-  def isRead(cmd: UInt): Bool                   = cmd(3)
-  def isWrite(cmd: UInt): Bool                  = !cmd(3)
-  def isMemRead(is_mem: Bool, cmd: UInt): Bool  = is_mem && cmd(3)
-  def isMemWrite(is_mem: Bool, cmd: UInt): Bool = is_mem && !cmd(3)
-}
-
-object RV32ILsuUtilities extends RegisteredLsuUtilities {
-  override def isaName: String     = "rv32i"
-  override def utils: LsuUtilities = new RV32ILsuUtilitiesImpl
+  override def factory: UtilitiesFactory[LsuUtilities] = LsuUtilitiesFactory
 }
